+++ src/DTO/ChiTietPhieuNhapDTO.java
@@ -1,28 +1,23 @@
 package DTO;
 
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-
 public class ChiTietPhieuNhapDTO {
     private int maphieunhap;
     private int masp;
     private String tensp;
     private int soluong;
-    private int dongia;
+    private double dongia;
 
     public ChiTietPhieuNhapDTO() {
     }
 
-    public ChiTietPhieuNhapDTO(int maphieunhap, int masp, int soluong, int dongia) {
+    public ChiTietPhieuNhapDTO(int maphieunhap, int masp, int soluong, double dongia) {
         this.maphieunhap = maphieunhap;
         this.masp = masp;
         this.soluong = soluong;
         this.dongia = dongia;
     }
     
-    public ChiTietPhieuNhapDTO(int masp, String tensp, int soluong, int dongia) {
+    public ChiTietPhieuNhapDTO(int masp, String tensp, int soluong, double dongia) {
         this.masp = masp;
         this.tensp = tensp;
         this.soluong = soluong;
@@ -61,11 +56,11 @@ public class ChiTietPhieuNhapDTO {
         this.soluong = soluong;
     }
 
-    public int getDongia() {
+    public double getDongia() {
         return dongia;
     }
 
-    public void setDongia(int dongia) {
+    public void setDongia(double dongia) {
         this.dongia = dongia;
     }
 }
